+++ client/src/test/scala/org/http4s/client/ClientRouteTestBattery.scala
@@ -8,13 +8,16 @@ import fs2._
 import org.http4s.Uri.{Authority, RegName}
 import org.http4s.client.testroutes.GetRoutes
 import org.http4s.dsl._
-
 import org.specs2.specification.core.Fragments
-import scala.concurrent.duration.FiniteDuration
+import scala.concurrent.duration._
+import scalaz.concurrent.Task
+import scalaz.stream.Process
 
 abstract class ClientRouteTestBattery(name: String, client: Client)
   extends Http4sSpec with JettyScaffold
 {
+  val timeout = 20.seconds
+
   Fragments.foreach(GetRoutes.getPaths.toSeq) { case (path, expected) =>
     s"Execute GET: $path" in {
       val name = address.getHostName